refactor: centralize backend URL helpers
